app/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
antlr4-python3-runtime==4.8
bitarray==2.5.1
certifi==2022.6.15
charset-normalizer==2.0.12
colorama==0.4.5
Cython==0.29.30
fairseq==0.12.2
hydra-core==1.0.7
idna==3.3
numpy==1.22.4
omegaconf==2.0.6
packaging==21.3
portalocker==2.4.0
pycparser==2.21
pyparsing==3.0.9
PyYAML==6.0
regex==2022.6.2
requests==2.28.0
sacrebleu==2.1.0
tabulate==0.8.9
torch==1.12.0+cpu
torchaudio==0.12.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
gunicorn==20.1.0
Flask==2.1.2
Flask-Cors==3.0.10
python-dotenv==0.20.0